fix: long compile times due to IntoView trait
This commit adds an explicit call to AnyView trait, which fixes long
compile times due to rustc having to "guess" the types of traits.
This would get insanely complex due to all the nested loops, taking a
couple minutes to "hot" reload.
With this change, hot reload times are back down to ~5 seconds.