feat: Enable caching of uploaded files
Makes the `/uploads/*` routes respond with appropriate headers to allow
clients to cache the requests:

- `last-modified`
- `etag`
- `cache-control`

while additionally supporting the `if-none-match` request header.

I'm not sure if the `last-modified` and `cache-control` headers are
strictly necessary to allow the client browser to cache the requests,
but adding them doesn't hurt and I don't have time to figure it out. :)

This required adding `entity_tag` as a new dependency to the backend
crate.

Resolves #23