feat: Add default slide group end date; Display dates in Stockholm time (#28)
* feat: Add default slide group end date

Makes the default slide group end date be the current time + one month,
instead of being infinite.

* feat: Display dates in Stockholm time

Displays the datetime inputs in Stockholm's current time zone, instead
of the user's local one.

Also removes the time zone offset from the formatted date time in the slide
group view.

This meant adding chrono-tz as a dependency for the frontend. This is a
pretty large library, but hopefully it's fine.

Also, I'm very tired now :3