fix(dev): Make `app` depend on `nyckeln`
Marks the `nyckeln` service as a dependency for the app service in the
Docker compose config. This makes `app` wait until `nyckeln` has started
before continuing. This seems to improve the issue I've personally been
having with the OIDC Manager failing to initialize when launching the
containers the first time after a code change. The issue was most
likely some sort of race condition (but it's very strange that it would
start successfully after restarting the containers).

This required adding a health checker function which is ran periodically
to determine if `nyckeln` is running. The test function isn't perfect as
it only checks that the service is listening on it's ports. I decided to
not have it try any actual endpoints, as that would result in the
requests being logged every 2 seconds.