feat(backend): Add endpoint to get Hive memberships
Adds an API endpoint to get the currently logged in user's Hive
memberships.

This required adding a new Hive client abstraction, as well as adding
support for configuring the Hive API token. Note: I have not tested the
changes to the production compose file (I wouldn't know how to). I've
tried to copy the pattern from the OIDC configuration.