feat: Add support for Hive group ownership (#40)
* style: Remove unused import in slide_group_options.rs

Removes unused import of `leptos::leptos_dom::logging::console_log` that
I accidentally left behind previously.

* chore: Add more test users to dev environment

Adds 3 new users to the development docker compose environment:
- Diana Datalog
- Martin Median
- Karin Klubbare

Also adds the group METAdorerna to that environment, and adds all users
to that group.

This will be used when developing the auth features.

* chore(backend): Add reqwest dependency

Adds the reqwest package as a dependency to the backend crate. This will
be used for the Hive client.

* chore: Add missing libs to shell.nix

Adds pkg-config and openssl to the Nix shell configuration. Not sure
why it wasn't a problem before, but I'm not able to run `cargo check`
without it.

Also added some environment variables which I'm like 80 % are necessary.

* fix(dev): Make `app` depend on `nyckeln`

Marks the `nyckeln` service as a dependency for the app service in the
Docker compose config. This makes `app` wait until `nyckeln` has started
before continuing. This seems to improve the issue I've personally been
having with the OIDC Manager failing to initialize when launching the
containers the first time after a code change. The issue was most
likely some sort of race condition (but it's very strange that it would
start successfully after restarting the containers).

This required adding a health checker function which is ran periodically
to determine if `nyckeln` is running. The test function isn't perfect as
it only checks that the service is listening on it's ports. I decided to
not have it try any actual endpoints, as that would result in the
requests being logged every 2 seconds.

* feat(backend): Add endpoint to get Hive memberships

Adds an API endpoint to get the currently logged in user's Hive
memberships.

This required adding a new Hive client abstraction, as well as adding
support for configuring the Hive API token. Note: I have not tested the
changes to the production compose file (I wouldn't know how to). I've
tried to copy the pattern from the OIDC configuration.

* feat: Add support for Hive group ownership

Adds support for storing slide group ownership as a Hive group in the
backend. This also required changing the frontend so that it can display
group ownership.

* feat(frontend): Support transfering ownership

Adds a new transfer ownership button to the slide group options view.